requirements.txt
fastapi
uvicorn
python-binance
python-dotenv
pandas
numpy
ta
websockets 